Extract methods of Community Blogs to module
Also extract spec to separate file
