+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-04-22 19:33:17 +0300 using RuboCop version 0.67.2.
+# on 2019-04-22 20:01:06 +0300 using RuboCop version 0.67.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -15,7 +15,7 @@ Lint/UriEscapeUnescape:
 Metrics/AbcSize:
   Max: 23
 
-# Offense count: 13
+# Offense count: 14
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
@@ -49,6 +49,7 @@ Naming/AccessorMethodName:
   Exclude:
     - 'lib/teamlab/modules/calendar.rb'
     - 'lib/teamlab/modules/community.rb'
+    - 'lib/teamlab/modules/community/community_blogs.rb'
     - 'lib/teamlab/modules/community/community_bookmarks.rb'
     - 'lib/teamlab/modules/crm.rb'
     - 'lib/teamlab/modules/files.rb'
